pyproject.toml
[project]
authors = [
  {name = "dreadew", email = "p74ur@yandex.ru"},
]
dependencies = [
  "anyio (==4.10.0)",
  "asyncpg (==0.30.0)",
  "backports-tarfile (==1.2.0)",
  "bcrypt (==4.3.0)",
  "build (==1.3.0)",
  "cachecontrol (==0.14.3)",
  "certifi (==2025.8.3)",
  "cffi (==2.0.0)",
  "charset-normalizer (==3.4.3)",
  "cleo (==2.1.0)",
  "crashtest (==0.4.1)",
  "distlib (==0.4.0)",
  "duckdb (==1.4.0)",
  "dulwich (==0.24.1)",
  "faker (==37.8.0)",
  "fastjsonschema (==2.21.2)",
  "filelock (==3.19.1)",
  "findpython (==0.7.0)",
  "greenlet (==3.2.4)",
  "h11 (==0.16.0)",
  "httpcore (==1.0.9)",
  "httpx (==0.28.1)",
  "idna (==3.10)",
  "importlib-metadata (==8.7.0)",
  "installer (==0.7.0)",
  "jaraco-classes (==3.4.0)",
  "jaraco-context (==6.0.1)",
  "jaraco-functools (==4.3.0)",
  "keyring (==25.6.0)",
  "more-itertools (==10.8.0)",
  "msgpack (==1.1.1)",
  "numpy (==2.3.3)",
  "packaging (==25.0)",
  "pandas (==2.3.2)",
  "pbs-installer (==2025.9.2)",
  "pkginfo (==1.12.1.2)",
  "platformdirs (==4.4.0)",
  "poetry (==2.2.0)",
  "poetry-core (==2.2.0)",
  "pycparser (==2.23)",
  "pyproject-hooks (==1.2.0)",
  "python-dateutil (==2.9.0.post0)",
  "python-dotenv (==1.1.1)",
  "pytz (==2025.2)",
  "rapidfuzz (==3.14.1)",
  "requests (==2.32.5)",
  "requests-toolbelt (==1.0.0)",
  "shellingham (==1.5.4)",
  "six (==1.17.0)",
  "sniffio (==1.3.1)",
  "sqlalchemy (==2.0.43)",
  "tomlkit (==0.13.3)",
  "trove-classifiers (==2025.9.11.17)",
  "typing-extensions (==4.15.0)",
  "tzdata (==2025.2)",
  "urllib3 (==2.5.0)",
  "virtualenv (==20.34.0)",
  "xattr (==1.2.0)",
  "zipp (==3.23.0)",
  "zstandard (==0.25.0)",
]
description = ""
license = {text = "MIT"}
name = "db-seeder"
readme = "README.md"
requires-python = ">=3.11,<4.0"
version = "0.1.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[dependency-groups]
dev = [
    "bandit (>=1.8.6,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]
